Move RefreshBuffer to Buffers::Refresh.
