+++ lib/vedeu/buffers/all.rb
@@ -20,5 +20,6 @@ end # Vedeu
 require 'vedeu/buffers/buffer'
 require 'vedeu/buffers/null'
 require 'vedeu/buffers/repository'
+require 'vedeu/buffers/refresh'
 require 'vedeu/buffers/virtual_buffer'
 require 'vedeu/buffers/virtual_buffers'
